Add :eyebrowse flag
Add :eyebrowse flag to display the buffer in an eyebrowse slot matching
a specific tag. Shackle switches to either the first existing slot
matching a tag, or creates a new slot if one does not exist.

The eyebrowse util functions might better belong in eyebrowse itself,
but I'll keep them here for now while I'm trying out the feature.

README is not updated.